requirements.txt
anyio==3.6.1
Authlib==1.0.1
black==22.6.0
certifi==2022.6.15
cffi==1.15.1
click==8.1.3
cryptography==37.0.4
fastapi==0.79.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
itsdangerous==2.1.2
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pycparser==2.21
pydantic==1.9.1
python-dotenv==0.20.0
rfc3986==1.5.0
sniffio==1.2.0
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.3.0
uvicorn==0.18.2
